Consolidate utility functions for HLSL tests
The HLSL tests had many interdependencies requiring unrelated tests
to be included in order for dependent parts of other tests to be run.

To eliminate these interdependencies, common utility functions
declared in DxcTestUtils.h are collected into a single cpp file
unimaginatively named DxcTestUtils.cpp.
