+++ tools/clang/unittests/HLSL/ValidationTest.cpp
@@ -29,93 +29,6 @@
 using namespace std;
 using namespace hlsl;
 
-void CheckOperationSucceeded(IDxcOperationResult *pResult, IDxcBlob **ppBlob) {
-  HRESULT status;
-  VERIFY_SUCCEEDED(pResult->GetStatus(&status));
-  VERIFY_SUCCEEDED(status);
-  VERIFY_SUCCEEDED(pResult->GetResult(ppBlob));
-}
-
-static bool CheckMsgs(llvm::StringRef text, llvm::ArrayRef<LPCSTR> pMsgs,
-                      bool bRegex) {
-  const char *pStart = !text.empty() ? text.begin() : nullptr;
-  const char *pEnd = !text.empty() ? text.end() : nullptr;
-  for (auto pMsg : pMsgs) {
-    if (bRegex) {
-      llvm::Regex RE(pMsg);
-      std::string reErrors;
-      VERIFY_IS_TRUE(RE.isValid(reErrors));
-      VERIFY_IS_TRUE(RE.match(text));
-    } else {
-      const char *pMatch = std::search(pStart, pEnd, pMsg, pMsg + strlen(pMsg));
-      if (pEnd == pMatch) {
-        WEX::Logging::Log::Comment(WEX::Common::String().Format(
-            L"Unable to find '%S' in text:\r\n%.*S", pMsg, (pEnd - pStart),
-            pStart));
-      }
-      VERIFY_ARE_NOT_EQUAL(pEnd, pMatch);
-    }
-  }
-  return true;
-}
-
-bool CheckMsgs(const LPCSTR pText, size_t TextCount, const LPCSTR *pErrorMsgs,
-               size_t errorMsgCount, bool bRegex) {
-  return CheckMsgs(llvm::StringRef(pText, TextCount),
-                   llvm::ArrayRef<LPCSTR>(pErrorMsgs, errorMsgCount), bRegex);
-}
-
-static
-bool CheckOperationResultMsgs(IDxcOperationResult *pResult,
-                              llvm::ArrayRef<LPCSTR> pErrorMsgs,
-                              bool maySucceedAnyway, bool bRegex) {
-  HRESULT status;
-  CComPtr<IDxcBlobEncoding> text;
-  VERIFY_SUCCEEDED(pResult->GetStatus(&status));
-  VERIFY_SUCCEEDED(pResult->GetErrorBuffer(&text));
-  const char *pStart = text ? (const char *)text->GetBufferPointer() : nullptr;
-  const char *pEnd = text ? pStart + text->GetBufferSize() : nullptr;
-  if (pErrorMsgs.empty() || (pErrorMsgs.size() == 1 && !pErrorMsgs[0])) {
-    if (FAILED(status) && pStart) {
-      WEX::Logging::Log::Comment(WEX::Common::String().Format(
-          L"Expected success but found errors\r\n%.*S", (pEnd - pStart),
-          pStart));
-    }
-    VERIFY_SUCCEEDED(status);
-  }
-  else {
-    if (SUCCEEDED(status) && maySucceedAnyway) {
-      return false;
-    }
-    CheckMsgs(llvm::StringRef((const char *)text->GetBufferPointer(),
-                              text->GetBufferSize()),
-              pErrorMsgs, bRegex);
-  }
-  return true;
-}
-
-bool CheckOperationResultMsgs(IDxcOperationResult *pResult,
-                              const LPCSTR *pErrorMsgs, size_t errorMsgCount,
-                              bool maySucceedAnyway, bool bRegex) {
-  return CheckOperationResultMsgs(
-      pResult, llvm::ArrayRef<LPCSTR>(pErrorMsgs, errorMsgCount),
-      maySucceedAnyway, bRegex);
-}
-
-std::string DisassembleProgram(dxc::DxcDllSupport &dllSupport,
-                               IDxcBlob *pProgram) {
-  CComPtr<IDxcCompiler> pCompiler;
-  CComPtr<IDxcBlobEncoding> pDisassembly;
-
-  if (!dllSupport.IsEnabled()) {
-    VERIFY_SUCCEEDED(dllSupport.Initialize());
-  }
-
-  VERIFY_SUCCEEDED(dllSupport.CreateInstance(CLSID_DxcCompiler, &pCompiler));
-  VERIFY_SUCCEEDED(pCompiler->Disassemble(pProgram, &pDisassembly));
-  return BlobToUtf8(pDisassembly);
-}
-
 class ValidationTest
 {
 public:
