+++ tools/clang/unittests/HLSL/VerifierTest.cpp
@@ -13,8 +13,6 @@
 #include <string>
 #include "CompilationResult.h"
 #include "HLSLTestData.h"
-#include "llvm/Support/ManagedStatic.h"
-#include "dxc/Support/HLSLOptions.h"
 
 #include <fstream>
 
@@ -22,9 +20,6 @@
 #include "HlslTestUtils.h"
 
 using namespace std;
-
-MODULE_SETUP(TestModuleSetup);
-MODULE_CLEANUP(TestModuleCleanup);
 
 // The test fixture.
 class VerifierTest
@@ -129,23 +124,6 @@ public:
     CheckVerifies(hlsl_test::GetPathToHlslDataFile(name).c_str());
   }
 };
-
-bool TestModuleSetup() {
-  // Use this module-level function to set up LLVM dependencies.
-  if (hlsl::options::initHlslOptTable()) {
-    return false;
-  }
-  return true;
-}
-
-bool TestModuleCleanup() {
-  // Use this module-level function to set up LLVM dependencies.
-  // In particular, clean up managed static allocations used by
-  // parsing options with the LLVM library.
-  ::hlsl::options::cleanupHlslOptTable();
-  ::llvm::llvm_shutdown();
-  return true;
-}
 
 TEST_F(VerifierTest, RunAttributes) {
   CheckVerifiesHLSL(L"attributes.hlsl");
